+++ src/main/kotlin/fi/aalto/cs/apluscourses/services/exercise/ExercisesUpdater.kt
@@ -1,6 +1,9 @@
 package fi.aalto.cs.apluscourses.services.exercise
 
-import com.intellij.openapi.components.*
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.components.State
+import com.intellij.openapi.components.Storage
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.platform.ide.progress.withBackgroundProgress
 import com.intellij.platform.util.progress.reportProgress